Reorganization of client src folder to combine actions, reducers and components
