+++ client/src/components/entries/utils.js
