+++ client/src/components/landing/user_signup.reducer.js
@@ -1,4 +1,4 @@
-import { ACCOUNT_PAGE_SUBMIT, PHONE_VERIFY_SUBMIT, CREATING_USER, SIGN_UP_ERROR, USER_CREATED, VERIFYING_CODE, CODE_VERIFIED, CODE_ERROR } from '../actions/user_signup';
+import { ACCOUNT_PAGE_SUBMIT, PHONE_VERIFY_SUBMIT, CREATING_USER, SIGN_UP_ERROR, USER_CREATED, VERIFYING_CODE, CODE_VERIFIED, CODE_ERROR } from './user_signup.actions.js';
 
 const initialState = {
   isCreatingUser: false,
