+++ client/src/components/trends/trends.actions.js
