+++ client/src/configureStore.js
@@ -3,7 +3,7 @@ import thunkMiddleware from 'redux-thunk';
 import { createLogger } from 'redux-logger';
 import createHistory from 'history/createHashHistory';
 import { routerMiddleware } from 'react-router-redux';
-import rootReducer from './reducers/rootReducer.js';
+import rootReducer from './rootReducer.js';
 
 
 const loggerMiddleware = createLogger();
@@ -18,8 +18,8 @@ const configureStore = (preloadedState) => {
   );
 
   if (module.hot) {
-    module.hot.accept('./reducers/rootReducer.js', () => {
-      const nextRootReducer = require('./reducers/rootReducer.js').default;
+    module.hot.accept('./rootReducer.js', () => {
+      const nextRootReducer = require('./rootReducer.js').default;
       store.replaceReducer(nextRootReducer);
     });
   }
